verto_ase: Add endpoint to fetch all questions for a specific quiz.
While the instructions state "The correct answer should not be included
in this response.", I am unsure if that means that the options shouldn't
be displayed, or if they should be displayed, just that there shouldn't
be an option marked as "correct".

I am interpreting it as display the options, but don't show which option
has been marked as correct, like when we take real quizzes.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>